// types.ts
export interface NavItem {
  label: string;
  href: string;
}

export interface CastClass {
  name: string;
  price: string;
  description: string;
  color: string;
}

export interface Review {
  name: string;
  age: string;
  content: string;
  rating: number;
}

export interface FAQItem {
  question: string;
  answer: string;
}